PutTaskDone 関数の実装

+++ tasks_model.go
@@ -3,8 +3,8 @@ package tomeit
 import "time"
 
 type Task struct {
-	id        int
-	userId    int
+	id        int64
+	userId    int64
 	name      string
 	priority  int
 	deadline  time.Time
